Add store (wip)
